Update Readme
A basic web application that I made in my database class, in which you can look up restaurants with filters on it and also food items.

To run it just upload the war file and create the database in your localhost and put the all the database information inside the Appconstants.java. Here is live version. http://104.236.252.94:8080/RestaurantDB/